setup webpack build for server
Having a separate webpack build for the server application is a result
of trying to prerender *styled* markup on the server.

The current problem is that, with universal rendering, styles are not
able to be "required" like they are on the client side thanks to
webpack.  As a result, we get a flood of "syntax errors" on the server
whenever node tries requiring css files.

One possible solution to this is inline styles. However, this approach
has the problem of bloating the markup string rendered on the server
with styles. This is exacerbated when a component is rendered multiple
times, as its styles will not be reused.

Another possibility is to simply disregard styles on the server and send
down unstyled markup that will then be styled by the client. However,
this would result in a FOUC (flash of unstyled content) until javascript
is loaded (and no styling at all if js is unavailable on the client).

So, what seems to currently be the best solution is to extract styles
into an external stylesheet, and have both the client and server
reference the classname identifiers output by webpack. To do this, we
need to have webpack compile the server app.